+++ src/Db/Mysql/Drivers/Mysqli/Driver.php
@@ -111,7 +111,7 @@ class Driver extends MysqlBase
     public function open(): bool
     {
         $option = $this->option;
-        $this->instance = $instance = new \mysqli($option['host'] ?? '127.0.0.1', $option['username'], $option['password'], $option['database'], $option['port'] ?? 3306);
+        $this->instance = $instance = new \mysqli($option['host'] ?? '127.0.0.1', $option['username'], $option['password'], $option['database'], (int) ($option['port'] ?? 3306));
         $instance->set_charset($option['charset'] ?? 'utf8');
         $this->execInitSqls();
